% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almost_deserts_map.R
\name{almost_deserts_map}
\alias{almost_deserts_map}
\title{Plot maps of Brazilian "almost news deserts"}
\usage{
almost_deserts_map(aggregation = "municipalities", export = F)
}
\arguments{
\item{aggregation}{aggregation (character). Can be one out of 3 levels:
"municipalities", "states" or "regions" (which can also be written in
Portuguese, with no accents: "municipios:, "estados" and "regioes").
Default is set to \code{aggregation = "municipalities"}.}

\item{export}{logical. Should the map be exported as a .jpg file? Default is
set to no \code{export = F}.}
}
\value{
A \code{ggplot} object with a Brazil map filled according to the
  municipality being an "almost news desert".
}
\description{
Plot a map of Brazilian "almost news deserts" (municipalities
  with 1 or 2 news organizations recorded by Atlas da Noticia's research).
}
\details{
\code{almost_deserts_map} returns \code{ggplot} object with a Brazil
  map filled according to the city being an "almost news desert". Spatial
  data is extracted using the \href{https://github.com/ipeaGIT/geobr}{geobr
  package}.
}
\examples{

# Plot a map of Brazilian almost news deserts

almost_deserts_map()

# Plot and export a map showing the absolute n. of municipalities
# that are almost news deserts in each Brazilian region

almost_deserts_map(aggregation = "regions", export = TRUE)
}
\seealso{
\code{\link{news_deserts_map}} and \code{\link{n_orgs_100k_map}} for
  other functions that generates maps, and \code{\link{almost_deserts_map}}
  for a function to extract data on the "almost news deserts".
}
