% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almost_deserts_state.R
\name{almost_deserts_state}
\alias{almost_deserts_state}
\title{Retrieve aggregate number of Brazilian "almost news deserts"}
\usage{
almost_deserts_state(regions = F)
}
\arguments{
\item{regions}{logical. Define if the aggregate data should correspond to
states or regions. Default is set to return states' data - \code{regions =
F}.}
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item uf: abbreviation of the state's name (\code{if regions =
  F}). \item regiao: region's name. \item qtd_quase_desertos: number of
  almost news deserts municipalities in the state/region.}
}
\description{
Retrieve the aggregate number of Brazilian "almost news deserts"
  (municipalities with 1 or 2 news organizations recorded by News Atlas'
  research) by state or region.
}
\details{
\code{almost_deserts_state} returns a dataset containing information
  on the aggregate number of news deserts municipalities in Brazilian states
  or regions.
}
\examples{

# Extract data on the aggregate number of "almost news deserts" municipalities in Brazilian states.

deserts_aggregate_state <- almost_deserts_state(regions = FALSE)

# Extract data on the aggregate number of "almost news deserts" municipalities in Brazilian regions.

deserts_aggregate_region <- almost_deserts_state(regions = TRUE)
}
