% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_orgs_100k.R
\name{n_orgs_100k}
\alias{n_orgs_100k}
\title{Retrieve data on the number of news organizations per 100k/inhabitants}
\usage{
n_orgs_100k()
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item municipio: municipality name. \item uf: abbreviation of the
  state's name. \item regiao: name of the country's region where the
  municipality is located at. \item qtd_veiculos: number of news
  organizations in the municipality. \item codmun: IBGE's (Brazilian
  Institute of Geography and Statistics) code for the municipality (7-digit).
  \item populacao: municipality's population. \item ano: year from IBGE's
  population records (note that data on news organizations collected by
  Atlas' team were last updated on Nov. 30th, 2019.) \item
  veiculos_por_100k_hab: number of news organizations per 100,000
  inhabitants. \item IDHM: Human Development Index for the municipality
  (Census 2010). \item IDHM_R: Human Development Index - \emph{per capita}
  income for the municipality (Census 2010). \item IDHM_E: Human Development
  Index - education for the municipality (Census 2010).}
}
\description{
Retrieve data on the number of news organizations per 100,000
  inhabitants in Brazilian municipalities.
}
\details{
\code{n_orgs_100k} returns a dataset containing information on
  municipalities with at least one news outlet recorded by Atlas da Noticia's
  research team up to November, 2019. It includes a variable on the number of
  organizations per 100,000 inhabitants. The function returns only those
  municipalities.
}
\examples{

# Extract data on all the municipalities with at least one news outlet.

municipalities_with_media <- n_orgs_100k()
}
