% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_orgs_100k_map.R
\name{n_orgs_100k_map}
\alias{n_orgs_100k_map}
\title{Plot Brazilian maps on the number of news organizations per 100k/inhabitants}
\usage{
n_orgs_100k_map(aggregation = "municipalities", export = F)
}
\arguments{
\item{aggregation}{aggregation (character). Can be one out of 3 levels:
"municipalities", "states" or "regions" (which can also be written in
Portuguese, with no accents: "municipios:, "estados" and "regioes").
Default is set to \code{aggregation = "municipalities"}.}

\item{export}{logical. Should the map be exported as a .jpg file? Default is
set to no \code{export = F}.}
}
\value{
A \code{ggplot} object with a Brazil map filled according to the
  number of news organizations per 100k/inhabitants.
}
\description{
Plot a map with information on the number of news organizations
  per 100k/inhabitants.
}
\details{
\code{n_orgs_100k_map} returns \code{ggplot} object with a Brazil
  map filled according to the number of news organizations per
  100k/inhabitants in the chosen aggregation level. Spatial data is extracted
  using the \href{https://github.com/ipeaGIT/geobr}{geobr package}.
}
\examples{

# Plot a map with the number of news organizations per
# 100k/inhabitants at the regional level.

n_orgs_100k_map(aggregation = "regions", export = FALSE)

# Save a ggplot object into an object
# called states_organizations_number with the number of news
# organizations per 100k/inhabitants in each of state

states_organizations_number <- n_orgs_100k_map(aggregation = "states", export = FALSE)

# Plot and export a map showing the number news organizations per
# 100k/inhabitants at the municipality level

n_orgs_100k_map(aggregation = "municipalities", export = TRUE)
}
\seealso{
\code{\link{news_deserts_map}} and \code{\link{almost_deserts_map}}
  for other functions that generates maps, and \code{\link{n_orgs_100k}} and
  \code{\link{n_orgs_100k_state}} for functions to extract data on the number
  of organizations per 100k/inhabitants.
}
