% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_orgs_100k_state.R
\name{n_orgs_100k_state}
\alias{n_orgs_100k_state}
\title{Retrieve aggregate number of news organizations per 100k/inhabitants}
\usage{
n_orgs_100k_state(regions = T)
}
\arguments{
\item{regions}{logical. Define if variables on the regional level should be
included in the dataset. Default is set to include a variable with data
aggregated to the regional level - \code{regions = T}. See details for more
information.}
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item estado: abbreviation of the state's name. \item regiao:
  region's name. \item med_idhm: mean of the Human Development Index at the
  municipalities of the corresponding state (Census 2010). \item med_idhm_e: mean of the
  Human Development Index - education at the municipalities of the
  corresponding state (Census 2010). \item med_idhm_r: mean of the Human Development Index -
  \emph{per capita} income at the municipalities of the corresponding state (Census 2010).
  \item populacao: states' population. \item qtd_veiculos: number of news
  organizations in the state. \item veiculos_por_100k_hab: number of news
  organizations per 100,000 inhabitants (state-level). \item ano: year from
  IBGE's (Brazilian Institute of Geography and Statistics) records - note
  that data on news organizations collected by Atlas' team were last updated
  on Nov. 30th, 2019. \item qtd_veiculos_regiao: number of news organizations
  in the region (\code{if regions = T}). \item populacao_regiao: regions'
  population (\code{if regions = T}). \item veiculos_por_100k_hab_regiao:
  number of news organizations per 100,000 inhabitants (regional-level)
  (\code{if regions = T}). }
}
\description{
Retrieve the aggregate number of news organizations per 100,000
  inhabitants in Brazilian municipalities by state or region.
}
\details{
\code{n_orgs_100k_state} returns a dataset containing information on
  the aggregate number of news deserts municipalities in Brazilian states or
  regions.
}
\examples{

# Extract data on the aggregate number of "news deserts" municipalities in Brazilian states.

deserts_aggregate_region_state <- n_orgs_100k_state(regions = FALSE)

# Extract data on the aggregate number of "news deserts", including Brazilian regions' variables.

deserts_aggregate_region <- n_orgs_100k_state(regions = TRUE)
}
