% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_deserts.R
\name{news_deserts}
\alias{news_deserts}
\title{Retrieve data on Brazilian "news deserts"}
\usage{
news_deserts()
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item codmun: IBGE's (Brazilian Institute of Geography and
  Statistics) code for the municipality (7-digit). \item municipio:
  municipality name. \item uf: abbreviation of the state's name. \item
  regiao: name of the country's region where the municipality is located at.
  \item populacao: municipality's population. \item ano: year from IBGE's population
  records (note that data on news organizations collected by Atlas' team were
  last updated on Nov. 30th, 2019.) \item IDHM: Human Development Index for
  the municipality (Census 2010). \item IDHM_R: Human Development Index - \emph{per capita}
  income for the municipality (Census 2010). \item IDHM_E: Human Development Index -
  education for the municipality (Census 2010).}
}
\description{
Retrieve data on Brazilian "news deserts": municipalities with
  no news organizations recorded by Atlas da Noticia's research.
}
\details{
\code{news_deserts} returns a dataset containing information on
  municipalities with no news organizations recorded by Atlas da Noticia's
  research team up to November, 2019 - which were designated as "news
  deserts". The function returns only those municipalities.
}
\examples{

# Extract data on all "news deserts" municipalities

deserts_dataset <- news_deserts()

}
