% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_deserts_map.R
\name{news_deserts_map}
\alias{news_deserts_map}
\title{Plot maps of Brazilian "news deserts"}
\usage{
news_deserts_map(aggregation = "municipalities",
                        percentage = F, export = F)
}
\arguments{
\item{aggregation}{aggregation (character). Can be one out of 3 levels:
"municipalities", "states" or "regions" (which can also be written in
Portuguese, with no accents: "municipios:, "estados" and "regioes").
Default is set to \code{aggregation = "municipalities"}.}

\item{percentage}{logical. Should the map be filled with absolute number of
news deserts municipalities or the percentage of them in the chosen
aggregation? Default is set to return the absolute number \code{percentage
= F}. If the user sets aggregation to the municipality level, this
parameter cannot be set to \code{percentage = T}.}

\item{export}{logical. Should the map be exported as a .jpg file? Default is
set to no \code{export = F}.}
}
\value{
A \code{ggplot} object with a Brazil map filled according to the
  number of news deserts.
}
\description{
Plot a map with information on the number of Brazilian "news
  deserts" (municipalities with no news organizations recorded by News Atlas'
  research).
}
\details{
\code{news_deserts_maps} returns \code{ggplot} object with a Brazil
  map filled according to the number of "news deserts". Spatial data is
  extracted using the \href{https://github.com/ipeaGIT/geobr}{geobr package}.
  Percentages are calculated based on the total number of municipalities in
  each state or region. When \code{aggregation = "states"}, data are
  categorized by quartiles.
}
\examples{

# Plot a map with the absolute number of news deserts in each Brazilian region

news_deserts_map(aggregation = "regions", percentage = FALSE, export = FALSE)

# Save a ggplot object into an object called states_deserts with the number of news
# deserts in each of them (in percentages)

states_deserts <- news_deserts_map(aggregation = "states", percentage = FALSE, export = FALSE)

# Plot and export a map showing the municipalities that are news deserts

news_deserts_map(aggregation = "municipalities", export = TRUE)
}
\seealso{
\code{\link{almost_deserts_map}} and \code{\link{n_orgs_100k_map}}
  for other functions that generates maps, and \code{\link{news_deserts}} and
  \code{\link{news_deserts_state}} for functions to extract data on "news
  deserts".
}
