% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_deserts_state.R
\name{news_deserts_state}
\alias{news_deserts_state}
\title{Retrieve aggregate number of Brazilian "news deserts"}
\usage{
news_deserts_state(regions = F)
}
\arguments{
\item{regions}{logical. Define if the aggregate data should correspond to
states or regions. Default is set to return states' data - \code{regions =
F}.}
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item uf: abbreviation of the state's name (\code{if regions =
  F}). \item regiao: region's name (\code{if regions = T}). \item
  qtd_cidades: number of municipalities in the state/region. \item
  qtd_desertos: number of news deserts municipalities in the state/region.}
}
\description{
Retrieve the aggregate number of Brazilian news deserts
  (municipalities with no news organizations recorded by News Atlas'
  research) by state or region.
}
\details{
\code{news_deserts_state} returns a dataset containing information on
  the aggregate number of news deserts municipalities in Brazilian states or
  regions.
}
\examples{

# Extract data on the aggregate number of "news deserts" municipalities in Brazilian states.

deserts_aggregate_region_state <- news_deserts_state(regions = FALSE)

# Extract data on the aggregate number of "news deserts" municipalities in Brazilian regions.

deserts_aggregate_region <- news_deserts_state(regions = TRUE)
}
