% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_region.R
\name{organizations_region}
\alias{organizations_region}
\title{Retrieve data on Brazilian news organizations by region}
\usage{
organizations_region(region, news = 1, media = "all")
}
\arguments{
\item{region}{the name of the Brazilian region for which the user wants to
retrieve data. User should choose only one region as input. Accepted inputs
are "All", "North", "Northeast", "Central-West", "Southeast" or "South". "All"
returns data for every news organization found by News Atlas. The user
should note that \code{region = "All"} takes longer to run.}

\item{news}{indicates if the organization is a news organization or no.
Default is set to \code{news = 1}, which retrieves only news organizations.
If \code{news = 0}, retrieves other organizations. And if \code{news =
"all"} retrieves all organizations.}

\item{media}{indicates the type of media used by the organization. Default
\code{media = "all"} includes all results. The other inputs allowed are
"print", "online", "radio" and "TV".}
}
\value{
A \code{data.frame} with the following variables:

  \itemize{ \item id: News Atlas' ID for the organization. \item
  nome_veiculo: news organization's name. \item media_source_id: News Atlas'
  data source. \item segment_id: media type ID. \item segmento: media type.
  \item city_id: News Atlas' internal city ID. \item municipio: municipality
  name. \item codmun: IBGE's (Brazilian Institute of Geography and
  Statistics) code for the municipality (7-digit). \item state_id: News
  Atlas' internal state ID. \item region: abbreviation of the state's name.
  \item region_id: News Atlas' internal region ID. \item regiao: name of the
  country's region. \item address: news organization address. \item
  annotations: any observations on the organization. \item email:
  organization's email address. \item employees_range_id: News Atlas'
  internal ID for the number of employees. \item num_funcionario: number of
  employees (categorical). \item periodicity: periodicity of the publication.
  \item eh_jornal: is it a news organization? "1" for yes. \item
  eh_site_pago: is there a paywall? "1" for yes, "0" for no ("0" includes all
  media). \item data_fechamento: shutdown date (in case the organization is
  no longer operating). \item ativo: is it active? "1" for yes, "0" for no.
  \item data_inclusao: date it was included into News Atlas' dataset. \item
  business_models: a \code{list}. Returns the organization business model.
  \item features: a \code{list}. Returns features of the organization. \item
  collaborators: a \code{list}. Returns collaborators who provided data about
  the organization. \item media_channels: a \code{list}. Returns the
  organizations' media channels (such as the website, Facebook page, Twitter
  profile and others). }
}
\description{
Retrieve detailed data on Brazilian news organizations, such as
  contact information, n. of employees, business models, media channels and
  others by the state where they are based.
}
\details{
\code{organizations_region} returns a detailed dataset on Brazilian
  news organizations, such as contact information, n. of employees, business
  models, media channels and others.
}
\examples{

# Extract data on all news organizations from the southeast region

municipalities_with_media <- organizations_region(region = "Southeast")

}
